+++ great_expectations/data_context/data_context_variables.py
@@ -366,11 +366,9 @@ class CloudDataContextVariables(DataContextVariables):
             )
 
     def _init_store(self) -> DataContextStore:
+        from great_expectations.data_context.cloud_constants import GeCloudRESTResource
         from great_expectations.data_context.store.data_context_store import (
             DataContextStore,
-        )
-        from great_expectations.data_context.store.ge_cloud_store_backend import (
-            GeCloudRESTResource,
         )
 
         store_backend: dict = {
