+++ great_expectations/data_context/store/ge_cloud_store_backend.py
@@ -1,14 +1,16 @@
 import json
 import logging
 from abc import ABCMeta
-from enum import Enum
 from typing import Any, Dict, List, Optional, Set, Tuple, Union, cast
 from urllib.parse import urljoin
 
 import requests
 
 from great_expectations.core.http import create_session
-from great_expectations.data_context.cloud_constants import CLOUD_DEFAULT_BASE_URL
+from great_expectations.data_context.cloud_constants import (
+    CLOUD_DEFAULT_BASE_URL,
+    GeCloudRESTResource,
+)
 from great_expectations.data_context.store.store_backend import StoreBackend
 from great_expectations.data_context.types.refs import GeCloudResourceRef
 from great_expectations.data_context.types.resource_identifiers import GeCloudIdentifier
@@ -107,30 +109,9 @@ def get_user_friendly_error_message(
     return " ".join(support_message)
 
 
-class GeCloudRESTResource(str, Enum):
-    BATCH = "batch"
-    CHECKPOINT = "checkpoint"
-    # Chetan - 20220811 - CONTRACT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-    CONTRACT = "contract"
-    DATASOURCE = "datasource"
-    DATA_ASSET = "data_asset"
-    DATA_CONTEXT = "data_context"
-    DATA_CONTEXT_VARIABLES = "data_context_variables"
-    EXPECTATION = "expectation"
-    EXPECTATION_SUITE = "expectation_suite"
-    EXPECTATION_VALIDATION_RESULT = "expectation_validation_result"
-    PROFILER = "profiler"
-    RENDERED_DATA_DOC = "rendered_data_doc"
-    # Chetan - 20220812 - SUITE_VALIDATION_RESULT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-    SUITE_VALIDATION_RESULT = "suite_validation_result"
-    VALIDATION_RESULT = "validation_result"
-
-
 class GeCloudStoreBackend(StoreBackend, metaclass=ABCMeta):
     PAYLOAD_ATTRIBUTES_KEYS: Dict[GeCloudRESTResource, str] = {
         GeCloudRESTResource.CHECKPOINT: "checkpoint_config",
-        # Chetan - 20220811 - CONTRACT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-        GeCloudRESTResource.CONTRACT: "checkpoint_config",
         GeCloudRESTResource.DATASOURCE: "datasource_config",
         GeCloudRESTResource.DATA_CONTEXT: "data_context_config",
         GeCloudRESTResource.DATA_CONTEXT_VARIABLES: "data_context_variables",
@@ -138,19 +119,12 @@ class GeCloudStoreBackend(StoreBackend, metaclass=ABCMeta):
         GeCloudRESTResource.EXPECTATION_VALIDATION_RESULT: "result",
         GeCloudRESTResource.PROFILER: "profiler",
         GeCloudRESTResource.RENDERED_DATA_DOC: "rendered_data_doc",
-        # Chetan - 20220812 - SUITE_VALIDATION_RESULT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-        GeCloudRESTResource.SUITE_VALIDATION_RESULT: "result",
         GeCloudRESTResource.VALIDATION_RESULT: "result",
     }
 
     ALLOWED_SET_KWARGS_BY_RESOURCE_TYPE: Dict[GeCloudRESTResource, Set[str]] = {
         GeCloudRESTResource.EXPECTATION_SUITE: {"clause_id"},
         GeCloudRESTResource.RENDERED_DATA_DOC: {"source_type", "source_id"},
-        # Chetan - 20220812 - SUITE_VALIDATION_RESULT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-        GeCloudRESTResource.SUITE_VALIDATION_RESULT: {
-            "checkpoint_id",
-            "expectation_suite_id",
-        },
         GeCloudRESTResource.VALIDATION_RESULT: {
             "checkpoint_id",
             "expectation_suite_id",
@@ -161,8 +135,6 @@ class GeCloudStoreBackend(StoreBackend, metaclass=ABCMeta):
         **{  # type: ignore[arg-type]
             GeCloudRESTResource.BATCH: "batches",
             GeCloudRESTResource.CHECKPOINT: "checkpoints",
-            # Chetan - 20220811 - CONTRACT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-            GeCloudRESTResource.CONTRACT: "contracts",
             GeCloudRESTResource.DATA_ASSET: "data_assets",
             GeCloudRESTResource.DATA_CONTEXT_VARIABLES: "data_context_variables",
             GeCloudRESTResource.DATASOURCE: "datasources",
@@ -171,8 +143,6 @@ class GeCloudStoreBackend(StoreBackend, metaclass=ABCMeta):
             GeCloudRESTResource.EXPECTATION_VALIDATION_RESULT: "expectation_validation_results",
             GeCloudRESTResource.PROFILER: "profilers",
             GeCloudRESTResource.RENDERED_DATA_DOC: "rendered_data_docs",
-            # Chetan - 20220812 - SUITE_VALIDATION_RESULT is deprecated by GX Cloud and is to be removed upon migration of E2E tests
-            GeCloudRESTResource.SUITE_VALIDATION_RESULT: "suite_validation_results",
             GeCloudRESTResource.VALIDATION_RESULT: "validation_results",
         }
     )
