+++ great_expectations/data_context/types/resource_identifiers.py
@@ -14,9 +14,7 @@ from great_expectations.core.run_identifier import RunIdentifier, RunIdentifierS
 from great_expectations.exceptions import DataContextError, InvalidDataContextKeyError
 
 if TYPE_CHECKING:
-    from great_expectations.data_context.store.ge_cloud_store_backend import (
-        GeCloudRESTResource,
-    )
+    from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 
 logger = logging.getLogger(__name__)
