+++ tests/data_context/cloud_data_context/test_checkpoint_crud.py
@@ -4,6 +4,7 @@ from unittest import mock
 
 import pytest
 
+from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 from great_expectations.data_context.data_context.abstract_data_context import (
     AbstractDataContext,
 )
@@ -14,9 +15,6 @@ from great_expectations.data_context.data_context.cloud_data_context import (
     CloudDataContext,
 )
 from great_expectations.data_context.data_context.data_context import DataContext
-from great_expectations.data_context.store.ge_cloud_store_backend import (
-    GeCloudRESTResource,
-)
 from great_expectations.data_context.types.base import (
     CheckpointConfig,
     DataContextConfig,
