+++ tests/data_context/cloud_data_context/test_expectation_suite_crud.py
@@ -4,11 +4,9 @@ from unittest import mock
 import pytest
 
 from great_expectations.core.expectation_suite import ExpectationSuite
+from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 from great_expectations.data_context.data_context.base_data_context import (
     BaseDataContext,
-)
-from great_expectations.data_context.store.ge_cloud_store_backend import (
-    GeCloudRESTResource,
 )
 from great_expectations.data_context.types.base import DataContextConfig, GeCloudConfig
 from great_expectations.data_context.types.resource_identifiers import GeCloudIdentifier
