+++ tests/data_context/store/conftest.py
@@ -1,10 +1,8 @@
 import pytest
 
 from great_expectations.core.serializer import JsonConfigSerializer
+from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 from great_expectations.data_context.store import DatasourceStore
-from great_expectations.data_context.store.ge_cloud_store_backend import (
-    GeCloudRESTResource,
-)
 from great_expectations.data_context.types.base import datasourceConfigSchema
