+++ tests/data_context/store/test_configuration_store.py
@@ -10,17 +10,8 @@ from ruamel.yaml.comments import CommentedMap
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.data_context_key import DataContextKey
+from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 from great_expectations.data_context.store import ConfigurationStore
-from great_expectations.data_context.store.ge_cloud_store_backend import (
-    GeCloudRESTResource,
-)
-from great_expectations.data_context.store.in_memory_store_backend import (
-    InMemoryStoreBackend,
-)
-from great_expectations.data_context.store.tuple_store_backend import (
-    TupleFilesystemStoreBackend,
-    TupleStoreBackend,
-)
 from great_expectations.data_context.types.base import BaseYamlConfig
 from great_expectations.data_context.types.resource_identifiers import (
     ConfigurationIdentifier,
