+++ tests/data_context/store/test_datasource_store_cloud_backend.py
@@ -4,10 +4,8 @@ from unittest import mock
 import pytest
 
 from great_expectations.core.serializer import DictConfigSerializer
+from great_expectations.data_context.cloud_constants import GeCloudRESTResource
 from great_expectations.data_context.store import DatasourceStore
-from great_expectations.data_context.store.ge_cloud_store_backend import (
-    GeCloudRESTResource,
-)
 from great_expectations.data_context.types.base import (
     DatasourceConfig,
     datasourceConfigSchema,
